refactor: changed typeorm entities setup
