--- packages/deployment-service/src/entities/abstract-entity.ts
@@ -1,12 +1,0 @@
-import { PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn } from 'typeorm'
-
-export abstract class AbsrtactEntity {
-  @PrimaryGeneratedColumn('uuid')
-  id?: string
-
-  @CreateDateColumn()
-  created?: string
-
-  @UpdateDateColumn()
-  modified?: string
-}
