+++ packages/deployment-service/src/executables/deploy-react.ts
@@ -1,5 +1,4 @@
-import { PipelineEntity } from './../entities/pipeline.entity'
-import { TaskEntity } from './../entities/task.entity'
+import { PipelineEntity, TaskEntity } from './../entities'
 import { ExecutableType } from './executable'
 
 export const deployReact: ExecutableType = (task: TaskEntity, pipeline: PipelineEntity): Promise<boolean> => {
