+++ packages/deployment-service/src/executables/executable.ts
@@ -1,5 +1,4 @@
-import { PipelineEntity } from './../entities/pipeline.entity'
-import { TaskEntity } from './../entities/task.entity'
+import { PipelineEntity, TaskEntity } from './../entities'
 import { TaskRunnerFunctions } from '@reapit/foundations-ts-definitions'
 import { build } from './build'
 import { deployNode } from './deploy-node'
