+++ packages/deployment-service/src/functions/pipeline/pipeline-get.ts
@@ -1,5 +1,5 @@
 import { httpHandler, NotFoundException } from '@homeservenow/serverless-aws-handler'
-import { PipelineEntity } from './../../entities/pipeline.entity'
+import { PipelineEntity } from './../../entities'
 import * as service from './../../services/pipeline'
 import { ownership, resolveDeveloperId } from './../../utils'
 import { defaultOutputHeaders } from './../../constants'
