+++ packages/deployment-service/src/functions/pipeline/pipeline-paginate.ts
@@ -1,5 +1,5 @@
 import { httpHandler } from '@homeservenow/serverless-aws-handler'
-import { PipelineEntity } from './../../entities/pipeline.entity'
+import { PipelineEntity } from './../../entities'
 import * as service from './../../services/pipeline'
 import { resolveDeveloperId } from './../../utils'
 import { Pagination } from 'nestjs-typeorm-paginate'
