+++ packages/deployment-service/src/functions/pipeline/pipeline-runner-create.ts
@@ -1,4 +1,4 @@
-import { PipelineEntity } from './../../entities/pipeline.entity'
+import { PipelineEntity } from './../../entities'
 import { ownership, resolveDeveloperId } from './../../utils'
 import { httpHandler, NotFoundException } from '@homeservenow/serverless-aws-handler'
 import * as service from '../../services'
