+++ packages/deployment-service/webpack.config.js
@@ -2,8 +2,14 @@ const slsw = require('serverless-webpack')
 const { webpackConfigNode, PATHS } = require('@reapit/ts-scripts')
 const CopyPlugin = require('copy-webpack-plugin')
 
+console.log(slsw)
+
 webpackConfigNode.entry = slsw.lib.entries
 webpackConfigNode.optimization.minimize = slsw.lib.webpack.isLocal ? false : true
+
+// new ResolveTSPathsToWebpackAlias({
+//   tsconfig: PATHS.tsConfig,
+// }),
 
 webpackConfigNode.plugins.push(new CopyPlugin({ 
   patterns: [
